Ya se muestra la informacion de la moneda en la nueva actividad, pero no hay una version exclusivamente funcional para el modo Landscape

+++ app/src/main/java/com/naldana/ejemplo10/models/Coin.kt
@@ -1,3 +1,3 @@
 package com.naldana.ejemplo10.models
 
-data class Coin (var name: String="Pedro coin", var country: String="", var value: Int=0, var values_us: Int=0, var year: Int=0, var isAvailable: Boolean=false, var img: String="")
+data class Coin (var name: String="Pedro coin", var country: String="", var value: Int=0, var values_us: Double = 0.0, var year: Int=0, var isAvailable: Boolean=false, var img: String="")
